requirements.txt
openai
streamlit
python-dotenv
pydub
